Move button actions into a dropdown rather than individual buttons
Previously, PRNFB would end up creating a separate button for each
action, which in turn would cause wrapping of the buttons and overall
not looking good, or being very usable if you had more than one or two
buttons.

This change moves them into a dropdown, making it a lot more compact
and a lot easier to add multiple actions.