Add Button Trigger Confirmation Dialog
When clicking the various trigger buttons, there is no feedback to the
user that the button was clicked. Even worse, the button may have been
clicked but there was an error doing the trigger itself (either in
the PRNFB code itself or when it actually does the final HTTP call
to the backing service).

This change adds an optional confirmation dialog (with a default of it
being disabled) that will report, after the button press is complete
and we have a response from the server, whether each trigger was
successful (or if no triggers were hit).