Add interactive forms to buttons
This change adds the ability to specify a JSON-based form for a given
button, which will get automatically rendered when the button is pressed.
The submitted data is available as serialized JSON in the ${BUTTON_FORM_DATA}
variable.

For the specification of what a form looks like and it's serialized result,
look at README.md in the change.