Change to force Bitbucket to perform merge
As per its current design, when you push updates to a pull request,
Bitbucket does not try and merge the updates until (usually) the user
visits the Pull Request UI page.

This means that if you notify, say, Jenkins to try and build your update
PR, it will in fact only build the last version that has been looked at
within the UI.

This change causes Bitbucket to try and do the merge prior to notifying
in order to have updated the Git references with the latest commits for
the external system to fetch and act on.